Extend PostingsForMatchers benchmark to cover Blocks too.
