+++ tsdb/head_bench_test.go
@@ -48,73 +48,3 @@ func BenchmarkHeadStripeSeriesCreateParallel(b *testing.B) {
 		}
 	})
 }
-
-func BenchmarkHeadPostingForMatchers(b *testing.B) {
-	h, err := NewHead(nil, nil, nil, 1000)
-	testutil.Ok(b, err)
-	defer func() {
-		testutil.Ok(b, h.Close())
-	}()
-
-	var ref uint64
-
-	addSeries := func(l labels.Labels) {
-		ref++
-		h.getOrCreateWithID(ref, l.Hash(), l)
-	}
-
-	for n := 0; n < 10; n++ {
-		for i := 0; i < 100000; i++ {
-			addSeries(labels.FromStrings("i", strconv.Itoa(i), "n", strconv.Itoa(n), "j", "foo"))
-			// Have some series that won't be matched, to properly test inverted matches.
-			addSeries(labels.FromStrings("i", strconv.Itoa(i), "n", strconv.Itoa(n), "j", "bar"))
-			addSeries(labels.FromStrings("i", strconv.Itoa(i), "n", "0_"+strconv.Itoa(n), "j", "bar"))
-			addSeries(labels.FromStrings("i", strconv.Itoa(i), "n", "1_"+strconv.Itoa(n), "j", "bar"))
-			addSeries(labels.FromStrings("i", strconv.Itoa(i), "n", "2_"+strconv.Itoa(n), "j", "foo"))
-		}
-	}
-
-	n1 := labels.MustNewMatcher(labels.MatchEqual, "n", "1")
-
-	jFoo := labels.MustNewMatcher(labels.MatchEqual, "j", "foo")
-	jNotFoo := labels.MustNewMatcher(labels.MatchNotEqual, "j", "foo")
-
-	iStar := labels.MustNewMatcher(labels.MatchRegexp, "i", "^.*$")
-	iPlus := labels.MustNewMatcher(labels.MatchRegexp, "i", "^.+$")
-	i1Plus := labels.MustNewMatcher(labels.MatchRegexp, "i", "^1.+$")
-	iEmptyRe := labels.MustNewMatcher(labels.MatchRegexp, "i", "^$")
-	iNotEmpty := labels.MustNewMatcher(labels.MatchNotEqual, "i", "")
-	iNot2 := labels.MustNewMatcher(labels.MatchNotEqual, "n", "2")
-	iNot2Star := labels.MustNewMatcher(labels.MatchNotRegexp, "i", "^2.*$")
-
-	cases := []struct {
-		name     string
-		matchers []*labels.Matcher
-	}{
-		{`n="1"`, []*labels.Matcher{n1}},
-		{`n="1",j="foo"`, []*labels.Matcher{n1, jFoo}},
-		{`j="foo",n="1"`, []*labels.Matcher{jFoo, n1}},
-		{`n="1",j!="foo"`, []*labels.Matcher{n1, jNotFoo}},
-		{`i=~".*"`, []*labels.Matcher{iStar}},
-		{`i=~".+"`, []*labels.Matcher{iPlus}},
-		{`i=~""`, []*labels.Matcher{iEmptyRe}},
-		{`i!=""`, []*labels.Matcher{iNotEmpty}},
-		{`n="1",i=~".*",j="foo"`, []*labels.Matcher{n1, iStar, jFoo}},
-		{`n="1",i=~".*",i!="2",j="foo"`, []*labels.Matcher{n1, iStar, iNot2, jFoo}},
-		{`n="1",i!=""`, []*labels.Matcher{n1, iNotEmpty}},
-		{`n="1",i!="",j="foo"`, []*labels.Matcher{n1, iNotEmpty, jFoo}},
-		{`n="1",i=~".+",j="foo"`, []*labels.Matcher{n1, iPlus, jFoo}},
-		{`n="1",i=~"1.+",j="foo"`, []*labels.Matcher{n1, i1Plus, jFoo}},
-		{`n="1",i=~".+",i!="2",j="foo"`, []*labels.Matcher{n1, iPlus, iNot2, jFoo}},
-		{`n="1",i=~".+",i!~"2.*",j="foo"`, []*labels.Matcher{n1, iPlus, iNot2Star, jFoo}},
-	}
-
-	for _, c := range cases {
-		b.Run(c.name, func(b *testing.B) {
-			for i := 0; i < b.N; i++ {
-				_, err := PostingsForMatchers(h.indexRange(0, 1000), c.matchers...)
-				testutil.Ok(b, err)
-			}
-		})
-	}
-}
